* 'master' of github.com:kete/kete: (65 commits)
  translation: The item type in the text 'All results in [item_type]' on the search page wasn't being translated. Also fixed a few capitalization issues.
  bugfix: fixing comment deletion issue, caused by nested set callback occuring before versioned callbacks, and moving decendants to the grandparent instead of deleting them all.
  bugfix: basket deletion wasn't working because taggings where left behind which triggered foreign key constraint issues.
  refinement: the topic type link on topic show pages should search for other topics of same privacy as the item.
  refinement: adding display of currently searched topic type to page title and page heading. Making topic type in search results more visible.
  refinement: floating the topic type name to the right, making it grey (so it's accessible but not obtrusive). Rewording 'Of type:' to 'About a:'
  bugfix: special characters in topic type names were casuing issues when searching by topic type.
  refinement: adding link back to parent topic on related item searches.
  bugfix: fixing issue that caused titles like ': Item Title' on non-topic items. In these cases, topic type shouldn't be output.
  refinement: switching to named scope for maintainability
  refinement: moving the topic type display and adding a system setting to turn it off.
  enhancement: ability to search topics by topic type. Linking the topic type value on topic show pages to the search.
  enhancement: show the topic type within search results. Only show the exact one, but has ability to display heirarchy (Topic -> Person)
  bugfix: validate_topic_type_entry action should require login, but not a site admin to access it.
  bugfix: extended field data was being lost when doing searches on choice results.
  translate: running new translation sort rake task on the English locale.
  translate: Adding new sort task, to order translation keys from a-z.
  translation: improving recent translation strings.
  enhancement: adding very basic topic type extended field value validation using ajax requests. green tick for valid, red tick for invalid. Should give a basic idea whether the format is acceptable or now.
  bugfix: previous commit didn't send the updated file, leaving some translation keys broken. Making sure it works this time.
  ...

+++ lib/tasks/upgrade.rake
@@ -25,7 +25,9 @@ namespace :kete do
                     'kete:upgrade:ensure_logins_all_valid',
                     'kete:upgrade:move_user_name_to_display_and_resolved_name',
                     'kete:upgrade:set_default_locale_for_existing_users',
-                    'kete:upgrade:add_basket_id_to_taggings']
+                    'kete:upgrade:add_basket_id_to_taggings',
+                    'kete:upgrade:make_baskets_private_notification_do_not_email',
+                    'kete:upgrade:add_nested_values_to_comments']
   namespace :upgrade do
     desc 'Privacy Controls require that Comment#commentable_private be set.  Update existing comments to have this data.'
     task :update_existing_comments_commentable_private => :environment do
@@ -251,10 +253,13 @@ namespace :kete do
         user_count += 1
       end
       # finally, lets removing the user name field mapping to prevent new user names from being set
-      content_type_id = ContentType.find_by_class_name('User').id
-      extended_field_id = ExtendedField.find_by_label('User Name').id
-      content_mapping = ContentTypeToFieldMapping.find_by_content_type_id_and_extended_field_id(content_type_id, extended_field_id)
-      content_mapping.destroy unless content_mapping.nil?
+      extended_field = ExtendedField.find_by_label('User Name')
+      if extended_field
+        content_type_id = ContentType.find_by_class_name('User').id
+        extended_field_id = extended_field.id
+        content_mapping = ContentTypeToFieldMapping.find_by_content_type_id_and_extended_field_id(content_type_id, extended_field_id)
+        content_mapping.destroy unless content_mapping.nil?
+      end
       p "#{user_count.to_s} users user_name moved to resolved_name" if user_count > 0
     end
 
@@ -333,6 +338,11 @@ namespace :kete do
       Basket.all.each do |basket|
         basket.settings[:private_item_notification] = 'do_not_email' if basket.settings[:private_item_notification].blank?
       end
+    end
+
+    desc "Add the parent_id, lft, and rgt values to comments that were created before acts_as_nested_set was put in place"
+    task :add_nested_values_to_comments => :environment do
+      Comment.renumber_all if (Comment.count(:conditions => { :lft => nil }) > 0)
     end
 
     desc 'Checks for mimetypes an adds them if needed.'
